Schematic board, cam files, and new new library
Files of the schematic and board with the cam files. and the new made
button.

As of now, I am having a lot of trouble adding the button to the
schematic. The button appears in the library in the board file, but not
in the schematic. I will try to figure this out in the next.